# satelite-catalogo-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CC:
    properties:
      id:
        type: string
      jefe:
        type: string
      nombre:
        type: string
      numero:
        type: integer
    type: object
  models.Line:
    properties:
      cantidad:
        type: integer
      comentario:
        type: string
      importe_linea:
        type: number
      numero_linea:
        type: integer
      product_id:
        type: string
      um:
        type: string
    type: object
...
  models.Solicitud:
    properties:
      aprobador:
        type: string
      cc:
        type: string
      description:
        type: string
      documents:
        items:
          type: string
        type: array
      fecha_contable:
        type: string
      fecha_solicitud:
        type: string
      id:
        type: string
      importe_total:
        type: number
      lines:
        items:
          $ref: '#/definitions/models.Line'
        type: array
      moneda:
        type: string
      nombre_solicitud:
        type: string
      solicitante:
        type: string
      state:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Catalogo service.
  title: Catalogo API
  version: "1.0"
paths:
...
  /cc/:
    get:
      description: Returns all Centros de Costo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CC'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List centros de costo
      tags:
      - cc
    post:
      consumes:
      - application/json
      description: Creates a new Centro de Costo
      parameters:
      - description: Centro de Costo info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CC'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create centro de costo
      tags:
      - cc
...
  /solicitud/:
    get:
      description: Returns solicitudes paginated
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List solicitudes paginated
      tags:
      - solicitudes
    post:
      consumes:
      - multipart/form-data
      description: Creates a new solicitud with optional files
      parameters:
      - description: Solicitud JSON
        in: formData
        name: solicitud
        required: true
        type: string
      - description: Attached files
        in: formData
        name: archivos
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create solicitud
      tags:
      - solicitudes
  /solicitud/{id}:
    delete:
      description: Deletes a solicitud by ID
      parameters:
      - description: Solicitud ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete solicitud
      tags:
      - solicitudes
    get:
      description: Returns a solicitud by its ID
      parameters:
      - description: Solicitud ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Solicitud'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get solicitud by ID
      tags:
      - solicitudes
    put:
      consumes:
      - application/json
      description: Updates an existing solicitud
      parameters:
      - description: Solicitud ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: payload
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Update solicitud
      tags:
      - solicitudes
  /solicitud/aprobar:
    get:
      description: Returns solicitudes for approval filtered by supervisor
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: State
        in: query
        name: state
        type: string
      - description: Solicitud ID
        in: query
        name: id
        type: string
      - description: Centro de costo
        in: query
        name: cc
        type: string
      - description: Fecha inicio
        in: query
        name: fechaInicio
        type: string
      - description: Fecha fin
        in: query
        name: fechaFin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List solicitudes to approve
      tags:
      - solicitudes
  /solicitud/filtradas:
    get:
      description: Returns solicitudes using filters and pagination
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: State
        in: query
        name: state
        type: string
      - description: Solicitud ID
        in: query
        name: id
        type: string
      - description: Fecha inicio
        in: query
        name: fechaInicio
        type: string
      - description: Fecha fin
        in: query
        name: fechaFin
        type: string
      - collectionFormat: csv
        description: Centros de costo
        in: query
        items:
          type: string
        name: ccs
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List solicitudes filtered paginated
      tags:
      - solicitudes
...
swagger: "2.0"